+++ integration/tests/basic/erc/conftest.py
@@ -11,7 +11,6 @@ from spl.token.instructions import (
     get_associated_token_address,
 )
 
-from utils.consts import REMAPPING_ZEPPELIN
 from utils.erc721ForMetaplex import ERC721ForMetaplex
 from utils.web3client import NeonChainWeb3Client
 
@@ -19,21 +18,8 @@ from utils.web3client import NeonChainWeb3Client
 @pytest.fixture(scope="function")
 def solana_associated_token_mintable_erc20(
     erc20_spl_mintable, sol_client, solana_account: Keypair
-) -> Generator[Keypair, None, None]:
-    token_mint = erc20_spl_mintable.token_mint_pubkey
-    trx = Transaction()
-    trx.add(create_associated_token_account(solana_account.pubkey(), solana_account.pubkey(), token_mint))
-    opts = TxOpts(skip_preflight=True, skip_confirmation=False)
-    sol_client.send_transaction(trx, solana_account, opts=opts)
-    solana_address = get_associated_token_address(solana_account.pubkey(), token_mint)
-    yield solana_account, token_mint, solana_address
-
-
-@pytest.fixture(scope="function")
-def solana_associated_token_mintable_erc20_new(
-    erc20_spl_mintable_new, sol_client, solana_account: Keypair
 ) -> Generator[tuple[Keypair, Pubkey, Pubkey], None, None]:
-    token_mint = erc20_spl_mintable_new.token_mint_pubkey
+    token_mint = erc20_spl_mintable.token_mint_pubkey
     trx = Transaction()
     trx.add(create_associated_token_account(solana_account.pubkey(), solana_account.pubkey(), token_mint))
     opts = TxOpts(skip_preflight=True, skip_confirmation=False)
@@ -47,19 +33,6 @@ def solana_associated_token_erc20(
     erc20_spl, sol_client, solana_account: Keypair
 ) -> Generator[tuple[Keypair, Pubkey, Pubkey], None, None]:
     token_mint = erc20_spl.token_mint.pubkey
-    trx = Transaction()
-    trx.add(create_associated_token_account(solana_account.pubkey(), solana_account.pubkey(), token_mint))
-    opts = TxOpts(skip_preflight=True, skip_confirmation=False)
-    sol_client.send_transaction(trx, solana_account, opts=opts)
-    solana_address = get_associated_token_address(solana_account.pubkey(), token_mint)
-    yield solana_account, token_mint, solana_address
-
-
-@pytest.fixture(scope="function")
-def solana_associated_token_erc20_new(
-    erc20_spl_new, sol_client, solana_account: Keypair
-) -> Generator[tuple[Keypair, Pubkey, Pubkey], None, None]:
-    token_mint = erc20_spl_new.token_mint.pubkey
     trx = Transaction()
     trx.add(create_associated_token_account(solana_account.pubkey(), solana_account.pubkey(), token_mint))
     opts = TxOpts(skip_preflight=True, skip_confirmation=False)
@@ -87,16 +60,3 @@ def invalid_nft_receiver(web3_client_session, faucet, accounts):
         "EIPs/ERC721/ERC721InvalidReceiver", "0.8.10", accounts[0], contract_name="ERC721Receiver"
     )
     return contract
-
-
-@pytest.fixture(scope="class")
-def multiple_actions_erc20_new(web3_client_session, accounts, erc20_spl_mintable_new):
-    contract, contract_deploy_tx = web3_client_session.deploy_and_get_contract(
-        "EIPs/ERC20/MultipleActionsNew",
-        "0.8.28",
-        accounts[0],
-        contract_name="MultipleActionsERC20New",
-        import_remapping=REMAPPING_ZEPPELIN,
-        constructor_args=["Test TTT", "TTT", 9],
-    )
-    return accounts[0], contract
