+++ integration/tests/basic/erc/test_ERC20SPL.py
@@ -4,19 +4,20 @@ import allure
 import pytest
 import web3
 from _pytest.config import Config
-from solana.rpc.commitment import Confirmed
-from solana.rpc.types import TokenAccountOpts
+from solana.rpc.types import TokenAccountOpts, TxOpts
 from solana.transaction import Transaction
 from solders.keypair import Keypair
 from solders.pubkey import Pubkey
 from spl.token import instructions
 from spl.token.constants import TOKEN_PROGRAM_ID
 
-from integration.tests.basic.helpers.rpc_checks import assert_solana_address_was_not_used_in_trx
+from integration.tests.basic.helpers.errors import ContractError
 from utils import metaplex
-from utils.consts import ZERO_ADDRESS, METAPLEX_ADDRESS, SPL_TOKEN_ADDRESS, CALL_SOLANA_ADDRESS, SOLANA_NATIVE_ADDRESS
+from utils.consts import ZERO_ADDRESS
 from utils.erc20wrapper import ERC20Wrapper
 from utils.helpers import gen_hash_of_block, wait_condition, create_invalid_address
+from utils.multiple_actions_wrapper import transfer_five_times
+
 from utils.web3client import NeonChainWeb3Client
 from utils.solana_client import SolanaClient
 from utils.accounts import EthAccounts
@@ -31,7 +32,7 @@ NO_ENOUGH_GAS_PARAMS = [
 
 
 @allure.feature("ERC Verifications")
-@allure.story("ERC20SPL: Tests for ERC20ForSPL contract")
+@allure.story("ERC20SPL: Tests for ERC20ForSPLNew contract")
 @pytest.mark.usefixtures("accounts", "web3_client", "sol_client")
 @pytest.mark.neon_only
 class TestERC20SPL:
@@ -41,15 +42,12 @@ class TestERC20SPL:
 
     @pytest.fixture(scope="class")
     def erc20_contract(self, erc20_spl, eth_bank_account, pytestconfig: Config) -> ERC20Wrapper:
-        if pytestconfig.getoption("--network") == "mainnet":
-            self.web3_client.send_neon(eth_bank_account, erc20_spl.account.address, 10)
         return erc20_spl
 
-    @pytest.fixture()
+    @pytest.fixture
     def restore_balance(self, erc20_contract):
         pass
 
-    @pytest.mark.mainnet
     def test_metaplex_data(self, erc20_contract):
         metaplex.wait_account_info(self.sol_client, erc20_contract.token_mint.pubkey)
         metadata = metaplex.get_metadata(self.sol_client, erc20_contract.token_mint.pubkey)
@@ -57,7 +55,6 @@ class TestERC20SPL:
         assert metadata["data"]["symbol"] == erc20_contract.symbol
         assert metadata["is_mutable"] is True
 
-    @pytest.mark.mainnet
     def test_balanceOf(self, erc20_contract):
         recipient_account = self.accounts[1]
         transfer_amount = random.randint(0, 1000)
@@ -80,7 +77,6 @@ class TestERC20SPL:
         symbol = erc20_contract.contract.functions.symbol().call()
         assert symbol == erc20_contract.symbol
 
-    @pytest.mark.mainnet
     def test_name(self, erc20_contract):
         name = erc20_contract.contract.functions.name().call()
         assert name == erc20_contract.name
@@ -101,26 +97,48 @@ class TestERC20SPL:
         self,
         erc20_contract,
     ):
-        with pytest.raises(
-            web3.exceptions.ContractLogicError,
-            match="burn amount exceeds balance",
-        ):
-            erc20_contract.burn(self.accounts[2], 1000)
+        burn_amount = 1000
+        account_balance = erc20_contract.get_balance(self.accounts[2].address)
+        with pytest.raises(web3.exceptions.ContractLogicError) as exc_info:
+            erc20_contract.burn(self.accounts[2], burn_amount)
+
+        error_expected = ContractError(
+            signature="ERC20InsufficientBalance(address, uint256, uint256)", arg_types=["address", "uint256", "uint256"]
+        )
+
+        assert error_expected.matches(
+            error_hex_to_match=exc_info.value.args[0]
+        ), f" Expected {error_expected.selector.hex()}, got {exc_info.value.args[0]}"
+
+        address, balance, needed = error_expected.decode_args(exc_info.value.args[0])
+        address_expected = self.accounts[2].address.lower()
+        assert address_expected == address, f"Expected address to be {address_expected}, but got {address}"
+        assert account_balance == balance, f"Expected balance to be {account_balance}, but got {balance}"
+        assert burn_amount == needed, f"Expected limit to be {0}, but got {needed}"
 
     def test_burn_more_than_total_supply(self, erc20_contract):
         total = erc20_contract.contract.functions.totalSupply().call()
-        with pytest.raises(
-            web3.exceptions.ContractLogicError,
-            match="burn amount exceeds balance",
-        ):
+        with pytest.raises(web3.exceptions.ContractLogicError) as exc_info:
             erc20_contract.burn(erc20_contract.account, total + 1)
+
+        error_expected = ContractError(
+            signature="ERC20InsufficientBalance(address, uint256, uint256)", arg_types=["address", "uint256", "uint256"]
+        )
+
+        assert error_expected.matches(
+            error_hex_to_match=exc_info.value.args[0]
+        ), f" Expected {error_expected.selector.hex()}, got {exc_info.value.args[0]}"
+
+        address, balance, needed = error_expected.decode_args(exc_info.value.args[0])
+        address_expected = erc20_contract.account.address.lower()
+        assert address_expected == address, f"Expected address to be {address_expected}, but got {address}"
+        assert needed == total + 1, f"Expected amount to be {total}, but got {balance}"
 
     @pytest.mark.parametrize("param, msg", NO_ENOUGH_GAS_PARAMS)
     def test_burn_no_enough_gas(self, erc20_contract, param, msg):
         with pytest.raises(ValueError, match=msg):
             erc20_contract.burn(erc20_contract.account, 1, **param)
 
-    @pytest.mark.mainnet
     def test_burnFrom(self, erc20_contract, restore_balance):
         new_account = self.accounts[1]
         balance_before = erc20_contract.contract.functions.balanceOf(erc20_contract.account.address).call()
@@ -136,21 +154,48 @@ class TestERC20SPL:
 
     def test_burnFrom_without_allowance(self, erc20_contract):
         new_account = self.accounts.create_account()
-        with pytest.raises(
-            web3.exceptions.ContractLogicError,
-            match="insufficient allowance",
-        ):
-            erc20_contract.burn_from(new_account, erc20_contract.account.address, 10)
+        allowance_expected = 0
+        burn_amount = 10
+        with pytest.raises(web3.exceptions.ContractLogicError) as exc_info:
+            erc20_contract.burn_from(new_account, erc20_contract.account.address, burn_amount)
 
-    def test_burnFrom_more_than_allowanced(self, erc20_contract):
+        error_expected = ContractError(
+            signature="ERC20InsufficientAllowance(address, uint256, uint256)",
+            arg_types=["address", "uint256", "uint256"],
+        )
+
+        assert error_expected.matches(
+            error_hex_to_match=exc_info.value.args[0]
+        ), f" Expected {error_expected.selector.hex()}, got {exc_info.value.args[0]}"
+
+        spender, allowance, needed = error_expected.decode_args(exc_info.value.args[0])
+        address_expected = new_account.address.lower()
+        assert address_expected == spender, f"Expected address to be {address_expected}, but got {spender}"
+        assert allowance_expected == allowance, f"Expected allowance to be {allowance_expected}, but got {allowance}"
+        assert needed == burn_amount, f"Expected amount to be {needed}, but got {burn_amount}"
+
+    def test_burnFrom_more_than_allowed(self, erc20_contract):
         new_account = self.accounts.create_account()
-        amount = 2
-        erc20_contract.approve(erc20_contract.account, new_account.address, amount)
-        with pytest.raises(
-            web3.exceptions.ContractLogicError,
-            match="insufficient allowance",
-        ):
-            erc20_contract.burn_from(new_account, erc20_contract.account.address, amount + 1)
+        allowance_expected = 2
+        burn_amount = allowance_expected + 1
+        erc20_contract.approve(erc20_contract.account, new_account.address, allowance_expected)
+        with pytest.raises(web3.exceptions.ContractLogicError) as exc_info:
+            erc20_contract.burn_from(new_account, erc20_contract.account.address, burn_amount)
+
+        error_expected = ContractError(
+            signature="ERC20InsufficientAllowance(address, uint256, uint256)",
+            arg_types=["address", "uint256", "uint256"],
+        )
+
+        assert error_expected.matches(
+            error_hex_to_match=exc_info.value.args[0]
+        ), f" Expected {error_expected.selector.hex()}, got {exc_info.value.args[0]}"
+
+        spender, allowance, needed = error_expected.decode_args(exc_info.value.args[0])
+        address_expected = new_account.address.lower()
+        assert address_expected == spender, f"Expected address to be {address_expected}, but got {spender}"
+        assert allowance_expected == allowance, f"Expected allowance to be {allowance_expected}, but got {allowance}"
+        assert needed == burn_amount, f"Expected amount to be {needed}, but got {burn_amount}"
 
     @pytest.mark.parametrize("param, msg", NO_ENOUGH_GAS_PARAMS)
     def test_burnFrom_no_enough_gas(self, erc20_contract, param, msg):
@@ -170,19 +215,21 @@ class TestERC20SPL:
         assert allowance == amount
 
     @pytest.mark.parametrize(
-        "block_len, expected_exception, msg",
+        "block_len, expected_exception",
         [
-            (
-                ZERO_ADDRESS,
-                web3.exceptions.ContractLogicError,
-                "approve to the zero address",
-            ),
+            (ZERO_ADDRESS, web3.exceptions.ContractLogicError),
         ],
     )
-    def test_approve_incorrect_address(self, erc20_contract, block_len, expected_exception, msg):
+    def test_approve_incorrect_address(self, erc20_contract, block_len, expected_exception):
         address = create_invalid_address(block_len) if isinstance(block_len, int) else block_len
-        with pytest.raises(expected_exception, match=msg):
+        with pytest.raises(expected_exception) as exc_info:
             erc20_contract.approve(erc20_contract.account, address, 1)
+
+        error_expected = ContractError(signature="ERC20InvalidSpender(address)", arg_types=["address"])
+
+        assert error_expected.matches(
+            error_hex_to_match=exc_info.value.args[0]
+        ), f" Expected {error_expected.selector.hex()}, got {exc_info.value.args[0]}"
 
     @pytest.mark.parametrize("param, msg", NO_ENOUGH_GAS_PARAMS)
     def test_approve_no_enough_gas(self, erc20_contract, param, msg):
@@ -215,45 +262,52 @@ class TestERC20SPL:
         assert balance_acc2_after == balance_acc2_before + amount
         assert total_before == total_after
 
-    def test_transfer_and_check_sol_account_list_is_correct(self, erc20_contract, restore_balance, evm_loader):
-        new_account = self.accounts.create_account()
-
-        receipt = erc20_contract.transfer(erc20_contract.account, new_account.address, 100)
-        precompiled_addresses = [METAPLEX_ADDRESS, SPL_TOKEN_ADDRESS, CALL_SOLANA_ADDRESS, SOLANA_NATIVE_ADDRESS]
-        for precompiled_address in precompiled_addresses:
-            program_address = evm_loader.ether2program(precompiled_address[2:])[0]
-            assert_solana_address_was_not_used_in_trx(
-                receipt["transactionHash"].hex(), program_address, self.web3_client, evm_loader
-            )
-
     @pytest.mark.parametrize(
-        "block_len, expected_exception, msg",
+        "block_len, expected_exception",
         [
             (
                 ZERO_ADDRESS,
                 web3.exceptions.ContractLogicError,
-                "transfer to the zero address",
             ),
         ],
     )
-    def test_transfer_incorrect_address(self, erc20_contract, block_len, expected_exception, msg):
+    def test_transfer_incorrect_address(self, erc20_contract, block_len, expected_exception):
         address = gen_hash_of_block(block_len) if isinstance(block_len, int) else block_len
-        with pytest.raises(expected_exception, match=msg):
+        with pytest.raises(expected_exception) as exc_info:
             erc20_contract.transfer(erc20_contract.account, address, 1)
 
+        error_expected = ContractError(signature="ERC20InvalidReceiver(address)", arg_types=["address"])
+
+        assert error_expected.matches(
+            error_hex_to_match=exc_info.value.args[0]
+        ), f" Expected {error_expected.selector.hex()}, got {exc_info.value.args[0]}"
+
     def test_transfer_more_than_balance(self, erc20_contract):
-        balance = erc20_contract.contract.functions.balanceOf(erc20_contract.account.address).call()
-        with pytest.raises(
-            web3.exceptions.ContractLogicError,
-            match="transfer amount exceeds balance",
-        ):
-            erc20_contract.transfer(erc20_contract.account, erc20_contract.account.address, balance + 1)
+        balance_expected = erc20_contract.contract.functions.balanceOf(erc20_contract.account.address).call()
+        transfer_amount = balance_expected + 1
+        with pytest.raises(web3.exceptions.ContractLogicError) as exc_info:
+            erc20_contract.transfer(erc20_contract.account, erc20_contract.account.address, transfer_amount)
+
+        error_expected = ContractError(
+            signature="ERC20InsufficientBalance(address, uint256, uint256)", arg_types=["address", "uint256", "uint256"]
+        )
+
+        assert error_expected.matches(
+            error_hex_to_match=exc_info.value.args[0]
+        ), f" Expected {error_expected.selector.hex()}, got {exc_info.value.args[0]}"
+
+        address, balance, needed = error_expected.decode_args(exc_info.value.args[0])
+        address_expected = erc20_contract.account.address.lower()
+        assert address_expected == address, f"Expected address to be {address_expected}, but got {address}"
+        assert balance == balance_expected, f"Expected balance {balance_expected}, got {balance}"
+        assert needed == transfer_amount, f"Expected amount to be {transfer_amount}, but got {balance}"
 
     @pytest.mark.parametrize("param, msg", NO_ENOUGH_GAS_PARAMS)
     def test_transfer_no_enough_gas(self, erc20_contract, param, msg):
         with pytest.raises(ValueError, match=msg):
             erc20_contract.transfer(erc20_contract.account, erc20_contract.account.address, 1, **param)
 
+    @pytest.mark.cost_report
     def test_transferFrom(self, erc20_contract, restore_balance):
         new_account = self.accounts.create_account()
         balance_acc1_before = erc20_contract.contract.functions.balanceOf(erc20_contract.account.address).call()
@@ -271,31 +325,57 @@ class TestERC20SPL:
 
     def test_transferFrom_without_allowance(self, erc20_contract):
         new_account = self.accounts.create_account()
-        with pytest.raises(
-            web3.exceptions.ContractLogicError,
-            match="insufficient allowance",  # InvalidAllowance error insufficient allowance
-        ):
+        allowance_expected = 0
+        transfer_amount = 10
+        with pytest.raises(web3.exceptions.ContractLogicError) as exc_info:
             erc20_contract.transfer_from(
                 signer=new_account,
                 address_from=erc20_contract.account.address,
                 address_to=new_account.address,
-                amount=10,
+                amount=transfer_amount,
+            )
+        error_expected = ContractError(
+            signature="ERC20InsufficientAllowance(address, uint256, uint256)",
+            arg_types=["address", "uint256", "uint256"],
+        )
+
+        assert error_expected.matches(
+            error_hex_to_match=exc_info.value.args[0]
+        ), f" Expected {error_expected.selector.hex()}, got {exc_info.value.args[0]}"
+
+        spender, allowance, needed = error_expected.decode_args(exc_info.value.args[0])
+        address_expected = new_account.address.lower()
+        assert address_expected == spender, f"Expected address to be {address_expected}, but got {spender}"
+        assert allowance_expected == allowance, f"Expected allowance to be {allowance_expected}, but got {allowance}"
+        assert needed == transfer_amount, f"Expected amount to be {needed}, but got {transfer_amount}"
+
+    def test_transferFrom_more_than_allowed(self, erc20_contract):
+        new_account = self.accounts.create_account()
+        allowance_expected = 2
+        transfer_amount = allowance_expected + 1
+        erc20_contract.approve(erc20_contract.account, new_account.address, allowance_expected)
+        with pytest.raises(web3.exceptions.ContractLogicError) as exc_info:
+            erc20_contract.transfer_from(
+                signer=new_account,
+                address_from=erc20_contract.account.address,
+                address_to=new_account.address,
+                amount=transfer_amount,
             )
 
-    def test_transferFrom_more_than_allowanced(self, erc20_contract):
-        new_account = self.accounts.create_account()
-        amount = 2
-        erc20_contract.approve(erc20_contract.account, new_account.address, amount)
-        with pytest.raises(
-            web3.exceptions.ContractLogicError,
-            match="insufficient allowance",
-        ):
-            erc20_contract.transfer_from(
-                signer=new_account,
-                address_from=erc20_contract.account.address,
-                address_to=new_account.address,
-                amount=amount + 1,
-            )
+        error_expected = ContractError(
+            signature="ERC20InsufficientAllowance(address, uint256, uint256)",
+            arg_types=["address", "uint256", "uint256"],
+        )
+
+        assert error_expected.matches(
+            error_hex_to_match=exc_info.value.args[0]
+        ), f" Expected {error_expected.selector.hex()}, got {exc_info.value.args[0]}"
+
+        spender, allowance, needed = error_expected.decode_args(exc_info.value.args[0])
+        address_expected = new_account.address.lower()
+        assert address_expected == spender, f"Expected address to be {address_expected}, but got {spender}"
+        assert allowance_expected == allowance, f"Expected allowance to be {allowance_expected}, but got {allowance}"
+        assert needed == transfer_amount, f"Expected amount to be {needed}, but got {transfer_amount}"
 
     def test_transferFrom_incorrect_address(self, erc20_contract):
         with pytest.raises(web3.exceptions.InvalidAddress):
@@ -306,20 +386,33 @@ class TestERC20SPL:
                 amount=1,
             )
 
+    # TODO Add EmptyAccount(bytes32) MissingMetaples(bytes32) InvalidTokenMint(bytes32) AmountExceedsUint64
     def test_transferFrom_more_than_balance(self, erc20_contract):
         new_account = self.accounts.create_account()
+        balance_expected = erc20_contract.contract.functions.balanceOf(erc20_contract.account.address).call()
         amount = erc20_contract.contract.functions.balanceOf(erc20_contract.account.address).call() + 1
         erc20_contract.approve(erc20_contract.account, new_account.address, amount)
-        with pytest.raises(
-            web3.exceptions.ContractLogicError,
-            match="transfer amount exceeds balance",
-        ):
+        with pytest.raises(web3.exceptions.ContractLogicError) as exc_info:
             erc20_contract.transfer_from(
                 signer=new_account,
                 address_from=erc20_contract.account.address,
                 address_to=new_account.address,
                 amount=amount,
             )
+
+        error_expected = ContractError(
+            signature="ERC20InsufficientBalance(address, uint256, uint256)", arg_types=["address", "uint256", "uint256"]
+        )
+
+        assert error_expected.matches(
+            error_hex_to_match=exc_info.value.args[0]
+        ), f" Expected {error_expected.selector.hex()}, got {exc_info.value.args[0]}"
+
+        address, balance, needed = error_expected.decode_args(exc_info.value.args[0])
+        address_expected = erc20_contract.account.address.lower()
+        assert address_expected == address, f"Expected address to be {address_expected}, but got {address}"
+        assert balance == balance_expected, f"Expected balance {balance_expected}, got {balance}"
+        assert needed == amount, f"Expected amount to be {amount}, but got {balance}"
 
     @pytest.mark.parametrize("param, msg", NO_ENOUGH_GAS_PARAMS)
     def test_transferFrom_no_enough_gas(self, erc20_contract, param, msg):
@@ -339,11 +432,28 @@ class TestERC20SPL:
         sol_balance_before = sol_client.get_balance(acc.pubkey()).value
         contract_balance_before = erc20_contract.contract.functions.balanceOf(erc20_contract.account.address).call()
 
+        opts = TokenAccountOpts(token_mint)
+        token_data = sol_client.get_token_accounts_by_owner_json_parsed(acc.pubkey(), opts).value[0]
+        token_balance_before = token_data.account.data.parsed["info"]["tokenAmount"]["amount"]
         erc20_contract.transfer_solana(erc20_contract.account, bytes(solana_address), amount)
+        wait_condition(
+            lambda: int(
+                sol_client.get_token_accounts_by_owner_json_parsed(acc.pubkey(), opts)
+                .value[0]
+                .account.data.parsed["info"]["tokenAmount"]["amount"]
+            )
+            > int(token_balance_before),
+            timeout_sec=30,
+        )
 
         sol_balance_after = sol_client.get_balance(acc.pubkey()).value
+        token_data = sol_client.get_token_accounts_by_owner_json_parsed(acc.pubkey(), opts).value[0]
+        token_balance_after = token_data.account.data.parsed["info"]["tokenAmount"]["amount"]
         contract_balance_after = erc20_contract.contract.functions.balanceOf(erc20_contract.account.address).call()
 
+        assert (
+            int(token_balance_after) - int(token_balance_before) == amount
+        ), "Token balance for sol account is not correct"
         assert contract_balance_before - contract_balance_after == amount, "Contract balance is not correct"
         assert sol_balance_after == sol_balance_before, "Sol balance is changed"
 
@@ -395,7 +505,7 @@ class TestERC20SPL:
                 )
             )
         )
-        sol_client.send_tx_and_check_status_ok(trx, acc)
+        sol_client.send_transaction(trx, acc, opts=TxOpts(skip_preflight=False, skip_confirmation=False))
 
         claim_amount = random.randint(10, sent_amount)
         erc20_contract.claim(erc20_contract.account, bytes(solana_address), claim_amount)
@@ -404,7 +514,11 @@ class TestERC20SPL:
         assert balance_after == balance_before - sent_amount + claim_amount, "Balance is not correct"
 
     def test_claimTo(
-        self, erc20_contract, sol_client, solana_associated_token_erc20: tuple[Keypair, Pubkey, Pubkey], pytestconfig
+        self,
+        erc20_contract,
+        sol_client,
+        solana_associated_token_erc20: tuple[Keypair, Pubkey, Pubkey],
+        pytestconfig,
     ):
         new_account = self.accounts.create_account()
         acc, token_mint, solana_address = solana_associated_token_erc20
@@ -429,7 +543,7 @@ class TestERC20SPL:
                 )
             )
         )
-        sol_client.send_tx_and_check_status_ok(trx, acc)
+        sol_client.send_transaction(trx, acc, opts=TxOpts(skip_preflight=False, skip_confirmation=False))
 
         claim_amount = random.randint(10, sent_amount)
         erc20_contract.claim_to(erc20_contract.account, bytes(solana_address), new_account.address, claim_amount)
@@ -493,36 +607,35 @@ class TestERC20SPLMintable:
         balance_after = erc20_contract.contract.functions.balanceOf(new_account.address).call()
         assert balance_after == amount
 
-    def test_mint_by_no_minter_role(self, erc20_contract):
-        recipient_account = self.accounts[1]
-        with pytest.raises(web3.exceptions.ContractLogicError, match="must have minter role to mint"):
-            erc20_contract.mint_tokens(recipient_account, recipient_account.address, 0)
-
     @pytest.mark.parametrize(
-        "address_to, expected_exception, msg",
+        "address_to, expected_exception",
         [
-            (
-                ZERO_ADDRESS,
-                web3.exceptions.ContractLogicError,
-                "mint to the zero address",
-            ),
+            (ZERO_ADDRESS, web3.exceptions.ContractLogicError),
         ],
     )
-    def test_mint_with_incorrect_address(self, erc20_contract, address_to, expected_exception, msg):
+    def test_mint_with_incorrect_address(self, erc20_contract, address_to, expected_exception):
         address_to = create_invalid_address(address_to) if isinstance(address_to, int) else address_to
-        with pytest.raises(expected_exception, match=msg):
+        with pytest.raises(expected_exception) as exc_info:
             erc20_contract.mint_tokens(erc20_contract.account, address_to, 10)
 
+        error_expected = ContractError(signature="ERC20InvalidReceiver(address)", arg_types=["address"])
+
+        assert error_expected.matches(
+            error_hex_to_match=exc_info.value.args[0]
+        ), f" Expected {error_expected.selector.hex()}, got {exc_info.value.args[0]}"
+
     def test_mint_with_too_big_amount(self, erc20_contract):
-        with pytest.raises(
-            web3.exceptions.ContractLogicError,
-            match="total mint amount exceeds uint64 max",
-        ):
+        with pytest.raises(web3.exceptions.ContractLogicError) as exc_info:
             erc20_contract.mint_tokens(
                 erc20_contract.account,
                 erc20_contract.account.address,
                 UINT64_LIMIT,
             )
+        error_expected = ContractError(signature="AmountExceedsUint64(uint256)", arg_types=["uint256"])
+
+        assert error_expected.matches(
+            error_hex_to_match=exc_info.value.args[0]
+        ), f" Expected {error_expected.selector.hex()}, got {exc_info.value.args[0]}"
 
     @pytest.mark.parametrize("param, msg", NO_ENOUGH_GAS_PARAMS)
     def test_mint_no_enough_gas(self, erc20_contract, param, msg):
@@ -548,11 +661,29 @@ class TestERC20SPLMintable:
         amount = random.randint(10000, 1000000)
         sol_balance_before = sol_client.get_balance(acc.pubkey()).value
         contract_balance_before = erc20_contract.contract.functions.balanceOf(erc20_contract.account.address).call()
+
+        opts = TokenAccountOpts(token_mint)
+        token_data = sol_client.get_token_accounts_by_owner_json_parsed(acc.pubkey(), opts).value[0]
+        token_balance_before = token_data.account.data.parsed["info"]["tokenAmount"]["amount"]
         erc20_contract.transfer_solana(erc20_contract.account, bytes(solana_address), amount)
+        wait_condition(
+            lambda: int(
+                sol_client.get_token_accounts_by_owner_json_parsed(acc.pubkey(), opts)
+                .value[0]
+                .account.data.parsed["info"]["tokenAmount"]["amount"]
+            )
+            > int(token_balance_before),
+            timeout_sec=30,
+        )
 
         sol_balance_after = sol_client.get_balance(acc.pubkey()).value
+        token_data = sol_client.get_token_accounts_by_owner_json_parsed(acc.pubkey(), opts).value[0]
+        token_balance_after = token_data.account.data.parsed["info"]["tokenAmount"]["amount"]
         contract_balance_after = erc20_contract.contract.functions.balanceOf(erc20_contract.account.address).call()
 
+        assert (
+            int(token_balance_after) - int(token_balance_before) == amount
+        ), "Token balance for sol account is not correct"
         assert contract_balance_before - contract_balance_after == amount, "Contract balance is not correct"
         assert sol_balance_after == sol_balance_before, "Sol balance is changed"
 
@@ -571,11 +702,7 @@ class TestERC20SPLMintable:
             lambda: len(sol_client.get_token_accounts_by_delegate_json_parsed(acc.pubkey(), opts).value) > 0,
             timeout_sec=30,
         )
-        token_account = (
-            sol_client.get_token_accounts_by_delegate_json_parsed(acc.pubkey(), opts, commitment=Confirmed)
-            .value[0]
-            .account
-        )
+        token_account = sol_client.get_token_accounts_by_delegate_json_parsed(acc.pubkey(), opts).value[0].account
         assert int(token_account.data.parsed["info"]["delegatedAmount"]["amount"]) == amount
         assert int(token_account.data.parsed["info"]["delegatedAmount"]["decimals"]) == erc20_contract.decimals
 
@@ -607,7 +734,7 @@ class TestERC20SPLMintable:
                 )
             )
         )
-        sol_client.send_tx_and_check_status_ok(trx, acc)
+        sol_client.send_transaction(trx, acc, opts=TxOpts(skip_preflight=False, skip_confirmation=False))
 
         claim_amount = random.randint(10, sent_amount)
         erc20_contract.claim(erc20_contract.account, bytes(solana_address), claim_amount)
@@ -645,7 +772,7 @@ class TestERC20SPLMintable:
                 )
             )
         )
-        sol_client.send_tx_and_check_status_ok(trx, acc)
+        sol_client.send_transaction(trx, acc, opts=TxOpts(skip_preflight=False, skip_confirmation=False))
 
         claim_amount = random.randint(10, sent_amount)
         erc20_contract.claim_to(erc20_contract.account, bytes(solana_address), new_account.address, claim_amount)
@@ -658,12 +785,10 @@ class TestERC20SPLMintable:
 
 @allure.feature("ERC Verifications")
 @allure.story("ERC20SPL: Tests for multiple actions in one transaction")
-@pytest.mark.usefixtures("accounts", "web3_client", "sol_client")
-@pytest.mark.neon_only
+@pytest.mark.usefixtures("web3_client", "accounts")
 class TestMultipleActionsForERC20:
     web3_client: NeonChainWeb3Client
     accounts: EthAccounts
-    sol_client: SolanaClient
 
     def test_mint_transfer_burn(self, multiple_actions_erc20):
         sender_account = self.accounts[0]
@@ -946,49 +1071,120 @@ class TestMultipleActionsForERC20:
         ), "Contract balance is not correct"
         assert user_balance == transfer_amount + user_balance_before, "User balance is not correct"
 
-    def test_parallel_trxs_transfer_read_balance_transfer(self, multiple_actions_erc20, faucet, solana_account):
+    @pytest.mark.cost_report
+    def test_transfer_five_times(self, multiple_actions_erc20):
         sender_account = self.accounts[0]
-        receiver_account = self.accounts[1]
         acc, contract = multiple_actions_erc20
-        trx_amount = 10
-        transfer_amount = 100
+        contract_balance_before = contract.functions.contractBalance().call()
+        user_balance_before = contract.functions.balance(acc.address).call()
+        mint_amount_0 = random.randint(10000, 100000000)
+        transfer_amount_1 = random.randint(1, mint_amount_0)
+        transfer_amount_2 = random.randint(1, mint_amount_0 - transfer_amount_1)
+        transfer_amount_3 = random.randint(1, mint_amount_0 - transfer_amount_1 - transfer_amount_2)
+        transfer_amount_4 = random.randint(1, mint_amount_0 - transfer_amount_1 - transfer_amount_2 - transfer_amount_3)
+        transfer_amount_5 = random.randint(
+            1, mint_amount_0 - transfer_amount_1 - transfer_amount_2 - transfer_amount_3 - transfer_amount_4
+        )
 
         tx = self.web3_client.make_raw_tx(sender_account)
-        instruction_tx = contract.functions.mint(transfer_amount * trx_amount * 2).build_transaction(tx)
+        instruction_tx = contract.functions.mint(mint_amount_0).build_transaction(tx)
         self.web3_client.send_transaction(sender_account, instruction_tx)
 
-        hashes = []
-        nonce = self.web3_client.get_nonce(sender_account.address)
-        for i in range(trx_amount):
-            tx = self.web3_client.make_raw_tx(sender_account, nonce=nonce + i)
-            transaction = contract.functions.transferReadBalanceTransfer(
-                transfer_amount, receiver_account.address
-            ).build_transaction(tx)
-            instruction_tx = self.web3_client._web3.eth.account.sign_transaction(transaction, sender_account.key)
-            signature = self.web3_client._web3.eth.send_raw_transaction(instruction_tx.rawTransaction)
-            hashes.append(signature.hex())
-        for tx_hash in hashes:
-            resp = self.web3_client.wait_for_transaction_receipt(tx_hash)
-            assert resp.status == 1, f"Transaction {tx_hash} failed"
+        transfer_five_times(
+            self.web3_client,
+            contract,
+            sender_account,
+            acc.address,
+            [transfer_amount_1, transfer_amount_2, transfer_amount_3, transfer_amount_4, transfer_amount_5],
+        )
+
+        contract_balance = contract.functions.contractBalance().call()
+        user_balance = contract.functions.balance(acc.address).call()
+
+        assert (
+            contract_balance
+            == contract_balance_before
+            + mint_amount_0
+            - transfer_amount_1
+            - transfer_amount_2
+            - transfer_amount_3
+            - transfer_amount_4
+            - transfer_amount_5
+        ), "Contract balance is not correct"
+        assert (
+            user_balance
+            == transfer_amount_1
+            + transfer_amount_2
+            + transfer_amount_3
+            + transfer_amount_4
+            + transfer_amount_5
+            + user_balance_before
+        ), "User balance is not correct"
 
 
-@pytest.fixture(scope="class")
-def new_factory_contract(web3_client, erc20_spl_mintable):
-    contract, tx = web3_client.deploy_and_get_contract(
-        "external/neon-contracts/ERC20ForSPL/contracts/test/ERC20ForSPLMintableFactoryV2",
-        "0.8.24",
-        erc20_spl_mintable.account,
-        contract_name="ERC20ForSPLMintableFactoryV2",
-    )
-    return contract
+@allure.feature("ERC Verifications")
+@allure.story("ERC20SPL: Tests for new ERC20ForSPL contract")
+class TestERC20SPLNewFeatures:
 
+    def test_solana_account_getter(self, erc20_spl_mintable, accounts):
+        acc = self.accounts[0]
+        solana_pubkey = erc20_spl_mintable.get_solana_account(acc.address)
 
-@pytest.fixture(scope="class")
-def new_token_contract(web3_client, erc20_spl_mintable):
-    contract, tx = web3_client.deploy_and_get_contract(
-        "external/neon-contracts/ERC20ForSPL/contracts/test/ERC20ForSPLMintableV2",
-        "0.8.24",
-        erc20_spl_mintable.account,
-        contract_name="ERC20ForSPLMintableV2",
-    )
-    return contract
+        assert isinstance(solana_pubkey, bytes), "Returned value is not bytes32"
+        assert len(solana_pubkey) == 32, "Invalid bytes32 length"
+
+    def test_get_account_delegate_data(self, erc20_spl_mintable, neon_user):
+
+        signer = erc20_spl_mintable.account
+        owner_address = erc20_spl_mintable.account.address
+        recipient_user = neon_user.neon_address
+        solana_acc = erc20_spl_mintable.get_solana_account(recipient_user)
+
+        amount_to_approve = random.randint(500, 1000)
+        erc20_spl_mintable.approve_solana(signer=signer, spender=solana_acc, amount=amount_to_approve)
+
+        delegate_address, delegate_amount = erc20_spl_mintable.get_account_delegate_data(owner_address)
+
+        assert isinstance(delegate_address, bytes), "Delegate address is not bytes32"
+        assert delegate_address != bytes(0), "Address has to be non zero"
+        assert len(delegate_address) == 32, "Invalid delegate address length"
+        assert isinstance(delegate_amount, int), "Delegated amount is not uint64"
+        assert delegate_address == solana_acc, f"Delegated address is expected to {solana_acc}"
+        assert delegate_amount == amount_to_approve, f"Expected delegation to be {amount_to_approve}"
+
+    def test_get_token_mint_ata(self, erc20_spl_mintable):
+        """Test getTokenMintATA(bytes32 account) returns a valid ATA address."""
+        solana_pubkey = erc20_spl_mintable.get_solana_account(erc20_spl_mintable.address)
+        ata_address = erc20_spl_mintable.get_token_mint_ata(solana_pubkey)
+
+        assert isinstance(ata_address, bytes), "Returned ATA is not bytes32"
+        assert len(ata_address) == 32, "Invalid ATA address length"
+
+    def test_transferSolanaFrom(
+        self,
+        erc20_spl_mintable,
+        neon_user,
+        evm_loader,
+        accounts,
+        solana_associated_token_mintable_erc20,
+        sol_client,
+    ):
+
+        acc_0 = accounts[0]
+        approve_amount = 700
+        transfer_amount = random.randint(1, 300)
+
+        solana_acc, token_mint, ata_address = solana_associated_token_mintable_erc20
+
+        erc20_spl_mintable.approve(erc20_spl_mintable.account, acc_0.address, approve_amount)
+
+        balance_before = erc20_spl_mintable.get_balance(erc20_spl_mintable.account.address)
+
+        trx = erc20_spl_mintable.transfer_solana_from(
+            acc_0, erc20_spl_mintable.account.address, bytes(ata_address), transfer_amount
+        )
+        assert trx.status == 1, f"trx: {trx} failed"
+
+        balance_after = erc20_spl_mintable.get_balance(erc20_spl_mintable.account.address)
+
+        assert balance_after == balance_before - transfer_amount
